+++ datanode/gateway/graphql/resolvers.go
@@ -2838,6 +2838,41 @@ func (r *mySubscriptionResolver) busEventsWithBatch(
 	}
 }
 
+func (r *mySubscriptionResolver) LiquidityProvisions(ctx context.Context, partyID *string, marketID *string) (<-chan []*vega.LiquidityProvision, error) {
+	req := &v2.ObserveLiquidityProvisionsRequest{
+		MarketId: marketID,
+		PartyId:  partyID,
+	}
+	stream, err := r.tradingDataClientV2.ObserveLiquidityProvisions(ctx, req)
+	if err != nil {
+		return nil, customErrorFromStatus(err)
+	}
+
+	c := make(chan []*types.LiquidityProvision)
+	go func() {
+		defer func() {
+			stream.CloseSend()
+			close(c)
+		}()
+		for {
+			received, err := stream.Recv()
+			if err == io.EOF {
+				r.log.Error("orders: stream closed by server", logging.Error(err))
+				break
+			}
+			if err != nil {
+				r.log.Error("orders: stream closed", logging.Error(err))
+				break
+			}
+			if lps := received.LiquidityProvisions; lps != nil {
+				c <- lps
+			}
+		}
+	}()
+
+	return c, nil
+}
+
 // START: Account Resolver
 
 type myAccountResolver VegaResolverRoot